Addresses comments on pr
